go secure server

+++ api/gameApi/game_api.go
@@ -9,7 +9,6 @@ import (
 	"github.com/gin-gonic/gin"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
-	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
 type GamesHandler struct {
@@ -22,86 +21,57 @@ func NewGamesHandler(db *db.Client) *GamesHandler {
 	}
 }
 
-// Returns a list of all the games
+// GetGames Returns a list of all the games
 func (h *GamesHandler) GetGames(ctx *gin.Context) {
-	var gquery types.GamesQuery
+	var games = make([]types.Game, 0)
 
-	if err := ctx.ShouldBindQuery(&gquery); err != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	filter := bson.D{}
-
-	if gquery.Username != "" {
-		filter = append(filter, bson.E{Key: "username", Value: gquery.Username})
-	}
-
-	if gquery.SearchFilter != "" {
-		filter = append(filter, bson.E{
-			Key: "title",
-			Value: bson.D{{
-				Key: "$regex",
-				Value: primitive.Regex{
-					Pattern: gquery.SearchFilter,
-					Options: "i",
-				},
-			}},
-		})
-	}
-
-	skip := uint32(gquery.Page-1) * gquery.ItemsPerPage
-	options := options.Find().
-		SetSkip(int64(skip)).
-		SetLimit(int64(gquery.ItemsPerPage))
-
-	var games []types.Game = make([]types.Game, 0)
+	filter := &bson.D{{Key: "username", Value: ctx.GetString("username")}}
 
 	// Querying the 'games' collection to retrieve all the documents
-	status, err := h.db.QueryCollection("games", &filter, options, &games)
+	status, err := h.db.QueryCollection("games", filter, nil, &games)
 	if err != nil {
-		ctx.JSON(status, gin.H{"error": err.Error()})
+		ctx.JSON(status, gin.H{"message": err.Error()})
 		return
 	}
 
 	ctx.IndentedJSON(status, games)
 }
 
-// Returns a game based on its ID
+// GetOneGame Returns a game based on its ID
 func (h *GamesHandler) GetOneGame(ctx *gin.Context, id string) {
 	var game []types.Game
 
 	// Converting the ID to an ObjectID
 	objectID, err := primitive.ObjectIDFromHex(id)
 	if err != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		ctx.JSON(http.StatusBadRequest, gin.H{"message": err.Error()})
 		return
 	}
 
 	// Querying the 'games' collection to retrieve the document with the provided ID
 	status, err := h.db.QueryCollection("games", &bson.D{{Key: "_id", Value: objectID}}, nil, &game)
 	if err != nil {
-		ctx.JSON(status, gin.H{"error": err.Error()})
+		ctx.JSON(status, gin.H{"message": err.Error()})
 		return
 	}
 
 	// Verifying if the game was found
 	if len(game) == 0 {
-		ctx.JSON(http.StatusNotFound, gin.H{"error": "Game not found"})
+		ctx.JSON(http.StatusNotFound, gin.H{"message": "Game not found"})
 		return
 	}
 
 	ctx.IndentedJSON(status, game[0])
 }
 
-// Adds a new Game to the database
+// AddGame Adds a new Game to the database
 // Error if the game already exists
-func (h *GamesHandler) AddGame(ctx *gin.Context) *types.Game {
+func (h *GamesHandler) AddGame(ctx *gin.Context) {
 	// Retrieving the game from the request body
 	var game types.Game
 	if err := ctx.ShouldBindJSON(&game); err != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return nil
+		ctx.JSON(http.StatusBadRequest, gin.H{"message": err.Error()})
+		return
 	}
 
 	// Declaring the inserting conditions for the document
@@ -109,65 +79,73 @@ func (h *GamesHandler) AddGame(ctx *gin.Context) *types.Game {
 		{Key: "title", Value: game.Title},
 	}
 
+	game.Username = ctx.GetString("username")
 	game.Date = time.Now().UTC().Format("2006-01-02T15:04:05.000Z")
 	game.Version = 1
 
 	id, status, err := h.db.InsertDocument("games", insertingConditions, game)
 	if err != nil {
-		ctx.JSON(status, gin.H{"error": err.Error()})
-		return nil
+		ctx.JSON(status, gin.H{"message": err.Error()})
+		return
 	}
 
 	// Checking if no ID was provided by the server
 	// In this case, the game already exists
 	if id == nil {
-		ctx.JSON(status, gin.H{"error": "Game already exists"})
-		return nil
+		ctx.JSON(status, gin.H{"message": "Game already exists"})
+		return
 	}
 
 	game.ID = id.(primitive.ObjectID)
 
-	return &game
+	ctx.JSON(status, game)
 }
 
-// Deletes a game based on its ID
+// DeleteGame Deletes a game based on its ID
 // Error if the game does not exist
-func (h *GamesHandler) DeleteGame(ctx *gin.Context, id string) string {
+func (h *GamesHandler) DeleteGame(ctx *gin.Context) {
+	// Checking if an ID was provided
+	id := ctx.Param("id")
+	if id == "" {
+		ctx.JSON(http.StatusBadRequest, gin.H{"message": "ID not provided"})
+		return
+	}
+
 	// Converting the ID to an ObjectID
 	objectID, err := primitive.ObjectIDFromHex(id)
 	if err != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return ""
+		ctx.JSON(http.StatusBadRequest, gin.H{"message": err.Error()})
+		return
 	}
 
 	status, err := h.db.DeleteDocument("games", &bson.D{{Key: "_id", Value: objectID}})
 	if err != nil {
-		ctx.JSON(status, gin.H{"error": err.Error()})
-		return ""
+		ctx.JSON(status, gin.H{"message": err.Error()})
+		return
 	}
 
-	return id
+	ctx.JSON(status, gin.H{"message": "Game deleted"})
 }
 
-// Edits a game based on its ID
+// EditGame Edits a game based on its ID
 // Error if the game does not exist
-func (h *GamesHandler) EditGame(ctx *gin.Context) *types.Game {
+func (h *GamesHandler) EditGame(ctx *gin.Context) {
 	// Retrieving the game from the request body
 	var game types.Game
 	if err := ctx.ShouldBindJSON(&game); err != nil {
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return nil
+		ctx.JSON(http.StatusBadRequest, gin.H{"message": err.Error()})
+		return
 	}
 
 	status, err := h.db.EditDocument("games", &bson.D{{Key: "_id", Value: game.ID}}, game)
 
 	if err != nil {
-		ctx.JSON(status, gin.H{"error": err.Error()})
-		return nil
+		ctx.JSON(status, gin.H{"message": err.Error()})
+		return
 	}
 
 	game.Date = time.Now().UTC().Format("2006-01-02T15:04:05.000Z")
 	game.Version++
 
-	return &game
+	ctx.JSON(status, game)
 }
